Рефакторит redux под redux-toolkit
