+++ package.json
@@ -37,6 +37,7 @@
     "node": ">=12.0.0"
   },
   "dependencies": {
+    "@reduxjs/toolkit": "^1.5.0",
     "axios": "0.21.1",
     "leaflet": "1.7.1",
     "prop-types": "15.7.2",
