+++ src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import App from './components/app';
 import {Provider} from 'react-redux';
 import {store} from './store';
-import {checkLogin} from "./store/user/api-actions";
+import {checkLogin} from "./store/user/slice";
 
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
