--- src/store/map/types.ts
@@ -1,14 +1,0 @@
-import Offer from "../../models/offer";
-
-export const UPDATE_HOVERED_OFFER = `map/updateHoveredOffer`;
-
-export interface UpdateHoveredOfferAction {
-  type: typeof UPDATE_HOVERED_OFFER,
-  payload: Offer
-}
-
-export interface MapState {
-  hoveredOffer: Offer
-}
-
-export type MapActionTypes = UpdateHoveredOfferAction;
