+++ src/store/redirect/index.ts
@@ -1,11 +1,7 @@
-import browserHistory from "../../services/browser-history";
-import {Dispatch, Store} from "redux";
-import {REDIRECT_TO_ROUTE} from "./types";
+import redirectReducer, {redirectMiddleware} from './slice';
 
-export const redirect = (_store: Store) => (next: Dispatch) => (action: any) => {
-  if (action.type === REDIRECT_TO_ROUTE) {
-    browserHistory.push(action.payload);
-  }
-
-  return next(action);
+export {
+  redirectMiddleware
 };
+
+export default redirectReducer;
